requirements.txt
asgiref==3.5.2
Django==4.1.2
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-tables2==2.4.1
django-tempus-dominus==5.1.2.17
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.4
sqlparse==0.4.3
